casted the `uint8_t` type integer to `uint64_t`

+++ ss_asm/make_code.cpp
@@ -5,5 +5,5 @@ uint64_t ss_asm::ss_asm::make_code(enum types t, uint64_t d)
     // uint64_t c = t << DATA_REGION;
     // c += d;
     // return c;
-    return (t << DATA_REGION_BITS) + d;
+    return (static_cast<uint64_t>(t) << DATA_REGION_BITS) + d;
 }
